+++ examples/demo-01/backend.tf
@@ -1,0 +1,11 @@
+#------------------------------------------------------------------------------
+# AWS Backend configuration
+#
+#------------------------------------------------------------------------------
+terraform {
+  backend "s3" {
+    bucket = "fdelacruz-tfstate"
+    key    = "terraform.tfstate"
+    region = "us-east-1"
+  }
+}