+++ examples/demo-01/data.tf
@@ -1,0 +1,3 @@
+data "aws_subnet_ids" "public_subnets" {
+  vpc_id = aws_vpc.main.id
+}