+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 * Initial component release.
 * Tested with Terraform `v0.13.0`.
 * Tested with AWS provider `v3.3.6`
-* Compute Environment creation feature.
-* Job queue creation feature.
-* Tasks queue creation feature.
+* Compute Environment creation capabilities.
+* Job queue creation capabilities.
+* Tasks definition creation capabilities.
+* Added demo (`demo-01`).
 * IAM roles, policies and attachments created.